requirements.txt
streamlit==1.31.0
tensorflow-cpu==2.15.0
numpy==1.26.4
pillow
h5py==3.10.0